+++ stm/rtc.h
@@ -1,0 +1,2 @@
+void rtc_init(void);
+mp_obj_t pyb_rtc_read(void);
